Re-track a project if LXD was removed and installed again
LXD project metadata is used to track workshop
projects. Not the ideal solution, needs relocating
to the state (it was there before workshop
cross-ported the state package). Hence, if LXD was
removed and installed back, the projects track
will be lost. This fix restart tracking for the
already existing ones.