Remove artificial dependency on lxd init
This simplifies the tutorial and GitHub Action at the cost of LXD not
working normally if a user installs Workshop and later wants to start
using LXD directly. The default LXD storage pool and network can still
be created using `lxd init` without `--auto` and accepting all the
default settings.